Use FastqSubmitter
